# old/experimental/api/R1/productOrder - deprecated/swagger.yaml
...
swagger: "2.0"
info:
  description: ""
  version: "0.1.1_inProgress"
  title: "API ProductOrder"
host: "api.orange.com"
basePath: "/"
schemes:
- "http"
consumes:
- "application/json;charset=utf-8"
produces:
- "application/json;charset=utf-8"
paths:
  /productOrder:
    post:
      tags:
      - "productOrder"
      operationId: "productOrderCreate"
      summary: "This operation creates a MEF productOrder"
      description: "This operation allows to create a MEF productOrder. \nReference\
        \ documents:\n- Carrier Ethernet Ordering Technical Specification - April\
        \ 2017\n- Ethernet Ordering Interface Profile Spec\n- Ethernet Ordering Framework\
        \ API Mapping\n\n\nAdditional business control are described in this operation\
        \ (work in progress):\n\nSpecific business errors for current operation will\
        \ be encapsulated in\n\nHTTP Response 422 Unprocessable entity\n\n - 100:\
        \ Missing order item (minimum 1)\n   At least one order item must be provided\n\
        \ - 101: Missing Buyer at order level\n   One and only one related party with\
        \ a \"Buyer\" role should be provided at the product order level.\n - 102:\
        \ A relatedParty is at the wrong level\n   The party role provided is not\
        \ managed - MEF allows to have\n\"Buyer\", \"Seller\", \"Billing Contact\"\
        , \"Order Contact\", \"Implementation Contact\", \"Technical Contact\" roles\
        \ at product order level\n\"UNI Site Contact\", \"UNI Alt Site Contact\",\
        \ \"ENNI Site Contact\", \"ENNI Alt Site Contact\" at product order item level.\n\
        \ - 103: Missing Buyer Contact at order level\n   One and only one related\
        \ party with a \"Order Contact\" role should be provided at the product order\
        \ level.\nBuyer Contact name & Telephone number must be provided.\n - 104:\
        \ Missing Buyer Implementation Contact at order level\n   One and only one\
        \ related party with a \"Implementation Contact\" role should be provided\
        \ at the product order level.\nImplementation Contact name & Telephone number\
        \ must be provided.\n - 105: Missing Buyer Technical contact at order level\n\
        \   One and only one related party with a \"Technical Contact\" role should\
        \ be provided at the product order level.\nTechnical Contact name, Telephone\
        \ number and email address must be provided.\n - 106: Addres information must\
        \ matches place type\n   If place type is 'formatted address' : addrLine1,\
        \ postCode, country must be there, addrLine2, locality, postCodeExtension,\
        \ stateOrProvince  could be there, other address information must not be filled\n\
        If place type is 'fielded address' : streetName, streetType, postCode must\
        \ be there, streetNr, streetNrSuffix, streetNrLast, streetSuffix, locality,\
        \ postCodeExtension, country, stateOrProvince could be there, other address\
        \ information must not be filled\n - 107: postCode extenstion needs postcode\n\
        \   A postCode extension must not be present whithout a postCode"
      deprecated: false
      parameters:
      - name: "ProductOrder"
        required: true
        in: "body"
        description: ""
        schema:
          $ref: "#/definitions/ProductOrderCreateIN"
      responses:
        201:
          description: "Success"
          schema:
            $ref: "#/definitions/ProductOrder"
        400:
          description: "Bad Request\n\nList of supported error codes:\n- 20: Invalid\
            \ URL parameter value\n- 21: Missing body\n- 22: Invalid body\n- 23: Missing\
            \ body field\n- 24: Invalid body field\n- 25: Missing header\n- 26: Invalid\
            \ header value\n- 27: Missing query-string parameter\n- 28: Invalid query-string\
            \ parameter value"
          schema:
            $ref: "#/definitions/ZError"
        401:
          description: "Unauthorized\n\nList of supported error codes:\n- 40: Missing\
            \ credentials\n- 41: Invalid credentials\n- 42: Expired credentials"
          schema:
            $ref: "#/definitions/ZError"
        403:
          description: "Forbidden\n\nList of supported error codes:\n- 50: Access\
            \ denied\n- 51: Forbidden requester\n- 52: Forbidden user\n- 53: Too many\
            \ requests"
          schema:
            $ref: "#/definitions/ZError"
        404:
          description: "Not Found\n\nList of supported error codes:\n- 60: Resource\
            \ not found"
          schema:
            $ref: "#/definitions/ZError"
        405:
          description: "Method Not Allowed\n\nList of supported error codes:\n- 61:\
            \ Method not allowed"
          schema:
            $ref: "#/definitions/ZError"
        408:
          description: "Request Time-out\n\nList of supported error codes:\n- 63:\
            \ Request time-out"
          schema:
            $ref: "#/definitions/ZError"
        422:
          description: "Unprocessable entity\n\nFunctional error\n\nSpecific encapsulated\
            \ business errors for current operation\n\n - 100: Missing order item\
            \ (minimum 1)\n   At least one order item must be provided\n - 101: Missing\
            \ Buyer at order level\n   One and only one related party with a \"Buyer\"\
            \ role should be provided at the product order level.\n - 102: A relatedParty\
            \ is at the wrong level\n   The party role provided is not managed - MEF\
            \ allows to have\n\"Buyer\", \"Seller\", \"Billing Contact\", \"Order\
            \ Contact\", \"Implementation Contact\", \"Technical Contact\" roles at\
            \ product order level\n\"UNI Site Contact\", \"UNI Alt Site Contact\"\
            , \"ENNI Site Contact\", \"ENNI Alt Site Contact\" at product order item\
            \ level.\n - 103: Missing Buyer Contact at order level\n   One and only\
            \ one related party with a \"Order Contact\" role should be provided at\
            \ the product order level.\nBuyer Contact name & Telephone number must\
            \ be provided.\n - 104: Missing Buyer Implementation Contact at order\
            \ level\n   One and only one related party with a \"Implementation Contact\"\
            \ role should be provided at the product order level.\nImplementation\
            \ Contact name & Telephone number must be provided.\n - 105: Missing Buyer\
            \ Technical contact at order level\n   One and only one related party\
            \ with a \"Technical Contact\" role should be provided at the product\
            \ order level.\nTechnical Contact name, Telephone number and email address\
            \ must be provided.\n - 106: Addres information must matches place type\n\
            \   If place type is 'formatted address' : addrLine1, postCode, country\
            \ must be there, addrLine2, locality, postCodeExtension, stateOrProvince\
            \  could be there, other address information must not be filled\nIf place\
            \ type is 'fielded address' : streetName, streetType, postCode must be\
            \ there, streetNr, streetNrSuffix, streetNrLast, streetSuffix, locality,\
            \ postCodeExtension, country, stateOrProvince could be there, other address\
            \ information must not be filled\n - 107: postCode extenstion needs postcode\n\
            \   A postCode extension must not be present whithout a postCode"
          schema:
            $ref: "#/definitions/ZError"
        500:
          description: "Internal Server Error\n\nList of supported error codes:\n\
            - 1: Internal error"
          schema:
            $ref: "#/definitions/ZError"
        503:
          description: "Service Unavailable\n\nList of supported error codes:\n- 5:\
            \ The service is temporarily unavailable\n- 6: Orange API is over capacity,\
            \ retry later !"
          schema:
            $ref: "#/definitions/ZError"
    get:
      tags:
      - "productOrder"
      operationId: "productOrderFind"
      summary: "This operation allows to search MEF productOrder"
      description: "This operation has not been designed with MEF specific search\
        \ constraints but standard search pattern applies.\n\nSpecific business errors\
        \ for current operation will be encapsulated in\n\nHTTP Response 422 Unprocessable\
        \ entity\n"
      deprecated: false
      parameters:
      - name: "externalId"
        required: false
        in: "query"
        description: ""
        type: "string"
      - name: "orderDate"
        required: false
        in: "query"
        description: ""
        type: "string"
      - name: "state"
        required: false
        in: "query"
        description: ""
        type: "string"
      - name: "requestedCompletionDate"
        required: false
        in: "query"
        description: ""
        type: "string"
      - name: "projectId"
        required: false
        in: "query"
        description: ""
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ProductOrderFindResp"
        400:
          description: "Bad Request\n\nList of supported error codes:\n- 20: Invalid\
            \ URL parameter value\n- 21: Missing body\n- 22: Invalid body\n- 23: Missing\
            \ body field\n- 24: Invalid body field\n- 25: Missing header\n- 26: Invalid\
            \ header value\n- 27: Missing query-string parameter\n- 28: Invalid query-string\
            \ parameter value"
          schema:
            $ref: "#/definitions/ZError"
        401:
          description: "Unauthorized\n\nList of supported error codes:\n- 40: Missing\
            \ credentials\n- 41: Invalid credentials\n- 42: Expired credentials"
          schema:
            $ref: "#/definitions/ZError"
        403:
          description: "Forbidden\n\nList of supported error codes:\n- 50: Access\
            \ denied\n- 51: Forbidden requester\n- 52: Forbidden user\n- 53: Too many\
            \ requests"
          schema:
            $ref: "#/definitions/ZError"
        404:
          description: "Not Found\n\nList of supported error codes:\n- 60: Resource\
            \ not found"
          schema:
            $ref: "#/definitions/ZError"
        405:
          description: "Method Not Allowed\n\nList of supported error codes:\n- 61:\
            \ Method not allowed"
          schema:
            $ref: "#/definitions/ZError"
        408:
          description: "Request Time-out\n\nList of supported error codes:\n- 63:\
            \ Request time-out"
          schema:
            $ref: "#/definitions/ZError"
        422:
          description: "Unprocessable entity\n\nFunctional error"
          schema:
            $ref: "#/definitions/ZError"
        500:
          description: "Internal Server Error\n\nList of supported error codes:\n\
            - 1: Internal error"
          schema:
            $ref: "#/definitions/ZError"
        503:
          description: "Service Unavailable\n\nList of supported error codes:\n- 5:\
            \ The service is temporarily unavailable\n- 6: Orange API is over capacity,\
            \ retry later !"
          schema:
            $ref: "#/definitions/ZError"
  /productOrder/{productOrderId}:
    get:
      tags:
      - "productOrder"
      operationId: "productOrderGet"
      summary: "This operation allows to retrieve one MEF productOrder from its id"
      description: "Complete description of a MEF productOrder\n\nSpecific business\
        \ errors for current operation will be encapsulated in\n\nHTTP Response 422\
        \ Unprocessable entity\n"
      deprecated: false
      parameters:
      - name: "productOrderId"
        in: "path"
        required: true
        type: "string"
        description: ""
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/ProductOrder"
        400:
          description: "Bad Request\n\nList of supported error codes:\n- 20: Invalid\
            \ URL parameter value\n- 21: Missing body\n- 22: Invalid body\n- 23: Missing\
            \ body field\n- 24: Invalid body field\n- 25: Missing header\n- 26: Invalid\
            \ header value\n- 27: Missing query-string parameter\n- 28: Invalid query-string\
            \ parameter value"
          schema:
            $ref: "#/definitions/ZError"
        401:
          description: "Unauthorized\n\nList of supported error codes:\n- 40: Missing\
            \ credentials\n- 41: Invalid credentials\n- 42: Expired credentials"
          schema:
            $ref: "#/definitions/ZError"
        403:
          description: "Forbidden\n\nList of supported error codes:\n- 50: Access\
            \ denied\n- 51: Forbidden requester\n- 52: Forbidden user\n- 53: Too many\
            \ requests"
          schema:
            $ref: "#/definitions/ZError"
        404:
          description: "Not Found\n\nList of supported error codes:\n- 60: Resource\
            \ not found"
          schema:
            $ref: "#/definitions/ZError"
        405:
          description: "Method Not Allowed\n\nList of supported error codes:\n- 61:\
            \ Method not allowed"
          schema:
            $ref: "#/definitions/ZError"
        408:
          description: "Request Time-out\n\nList of supported error codes:\n- 63:\
            \ Request time-out"
          schema:
            $ref: "#/definitions/ZError"
        422:
          description: "Unprocessable entity\n\nFunctional error"
          schema:
            $ref: "#/definitions/ZError"
        500:
          description: "Internal Server Error\n\nList of supported error codes:\n\
            - 1: Internal error"
          schema:
            $ref: "#/definitions/ZError"
        503:
          description: "Service Unavailable\n\nList of supported error codes:\n- 5:\
            \ The service is temporarily unavailable\n- 6: Orange API is over capacity,\
            \ retry later !"
          schema:
            $ref: "#/definitions/ZError"
definitions:
...
  ActionType:
    description: ""
    type: "string"
    enum:
    - "add"
    - "modify"
    - "noChange"
    - "delete"
  StateType:
    description: ""
    type: "string"
    enum:
    - "acknowledge"
    - "inProgress"
    - "cancelled"
    - "completed"
    - "rejected"
    - "pending"
    - "held"
    - "failed"
    - "partial"
...
  OrderActivity:
    description: ""
    type: "string"
    enum:
    - "install"
    - "change"
    - "disconnect"
  desiredResponse:
    description: ""
    type: "string"
    enum:
    - "Confirmation and Engineering"
    - "Confirmation"
    - "None"
  pricingMethod:
    description: ""
    type: "string"
    enum:
    - "Tariff"
    - "Contract"
    - "Individual Case Basis"
    - "Other"
  ZErrorDetail:
    description: ""
    required:
    - "code"
    - "message"
    type: "object"
    properties:
      code:
        description: ""
        type: "integer"
        format: "int32"
      message:
        description: ""
        type: "string"
      description:
        description: ""
        type: "string"
      infoURL:
        description: ""
        type: "string"
  ZError:
    description: ""
    required:
    - "code"
    - "message"
    type: "object"
    properties:
      code:
        description: ""
        type: "integer"
        format: "int32"
      message:
        description: ""
        type: "string"
      description:
        description: ""
        type: "string"
      infoURL:
        description: ""
        type: "string"
      details:
        description: ""
        type: "array"
        items:
          $ref: "#/definitions/ZErrorDetail"
  BillingAccountRef:
    description: ""
    required:
    - "id"
    type: "object"
    properties:
      id:
        description: "Unique identifier of the billing account"
        type: "string"
...
  ProductOrder:
    description: ""
    required:
    - "externalId"
    - "orderDate"
    - "requestedCompletionDate"
    - "orderActivity"
    - "desiredResponses"
    - "version"
    - "orderItem"
    type: "object"
    properties:
      id:
        description: "ID created on repository side (OM system)\nMEF: Unique identifier\
          \ for the order that is generated by the Seller when the order is initially\
          \ accepted via an API."
        type: "string"
      externalId:
        description: "ID given by the consumer and only understandable by him (to\
          \ facilitate his searches afterwards).\nMEF: Buyer Purchase Order Number"
        type: "string"
      orderDate:
        description: "Date when the order was created"
        type: "string"
        format: "date-time"
      state:
        description: "Tracks the lifecycle status of the product order, such as Acknowledged,\
          \ Rejected, InProgress, Pending and so on\nMEF: to be seen if Desired Response\
          \ could be managed through state (9)"
        $ref: "#/definitions/StateType"
      requestedStartDate:
        description: "Order start date wished by the requestor\nMEF: If specified,\
          \ identifies the Buyer's desired date for order processing to start."
        type: "string"
        format: "date-time"
      requestedCompletionDate:
        description: "Requested delivery date from the requestor perspective"
        type: "string"
        format: "date-time"
      completionDate:
        description: "Date when the order was completed."
        type: "string"
        format: "date-time"
      expeditePriority:
        description: "Indicates that expedited treatment is requested and any charges\
          \ generated in provisioning this request (e.g., additional engineering charges\
          \ or labor charges if applicable) will be accepted"
        type: "integer"
        format: "int32"
      orderActivity:
        description: "MEF: The type of order, as defined by the business (e.g., install,\
          \ change, disconnect)"
        $ref: "#/definitions/OrderActivity"
      desiredResponses:
        description: "MEF: The type of order responses that the buyer wishes to receive\
          \ from the seller"
        $ref: "#/definitions/desiredResponse"
      version:
        description: "MEF: The version number that the Buyer uses to refer to this\
          \ particular version of the order"
        type: "string"
      projectId:
        description: "An identifier that is used to group Orders that represent a\
          \ unit of functionality that is important to a Buyer. A Project can be used\
          \ to relate multiple Orders together."
        type: "string"
      tspRestorationPriority:
        description: "Within the United States, indicates the provisioning and restoration\
          \ priority as defined under the TSP Service Vendor Handbook"
        type: "string"
      relatedParty:
        description: ""
        type: "array"
        items:
          $ref: "#/definitions/RelatedParty"
      note:
        description: "Extra information about the product order"
        $ref: "#/definitions/Note"
      orderItem:
        description: "An identified part of the order. A product order is decomposed\
          \ into one or more order item"
        type: "array"
        items:
          $ref: "#/definitions/OrderItem"
      '@type':
        description: ""
        type: "string"
      '@schemaLocation':
        description: ""
        type: "string"
  Note:
    description: ""
    required:
    - "text"
    - "date"
    - "author"
    type: "object"
    properties:
      text:
        description: "Texte of the note"
        type: "string"
      date:
        description: "Date of the note"
        type: "string"
        format: "date-time"
      author:
        description: "Author of the note"
        type: "string"
  OrderItem:
    description: ""
    required:
    - "id"
    - "action"
    - "productOffering"
    type: "object"
    properties:
      id:
        description: "Identifier of the line item (generally it is a sequence number\
          \ 01, 02, 03, ...)"
        type: "string"
      action:
        description: "Can be \"add\" / \"modify\" / \"no_change\"/ \"delete\""
        $ref: "#/definitions/ActionType"
      state:
        description: "State of the order item : described in the state machine diagram"
        $ref: "#/definitions/StateType"
      billingAccount:
        description: "Identifier of the billing account where to bill the charges"
        $ref: "#/definitions/BillingAccountRef"
      productOffering:
        description: "A product offering reference (ProductOfferingRef). A product\
          \ offering represents entities that are orderable from the provider of the\
          \ catalog, this resource includes pricing information"
        $ref: "#/definitions/ProductOffering"
      product:
        description: ""
        $ref: "#/definitions/Product"
      orderItemRelationship:
        description: ""
        type: "array"
        items:
          $ref: "#/definitions/OrderItemRelationShip"
      qualificationRef:
        description: "Used to refer serviceability request done previously."
        type: "array"
        items:
          $ref: "#/definitions/RelatedQualification"
      '@type':
        description: ""
        type: "string"
      '@schemaLocation':
        description: ""
        type: "string"
      pricingMethod:
        description: "The pricing authority to be used to price the products on this\
          \ order. Choice of Tariff, Contract, Individual Case Basis (ICB), Other"
        $ref: "#/definitions/pricingMethod"
      pricingReference:
        description: "The value associated with the Pricing Method selected (e.g.\
          \ a contract or tariff id)"
        type: "string"
      pricingTerm:
        description: "The length of the contract in months"
        type: "integer"
        format: "int32"
  ProductOffering:
    description: ""
    required:
    - "id"
    type: "object"
    properties:
      id:
        description: "MEF: Promotion ID (39)"
        type: "string"
  Product:
    description: ""
    required: []
    type: "object"
    properties:
      id:
        description: ""
        type: "string"
      '@type':
        description: ""
        type: "string"
      productSpecification:
        description: ""
        $ref: "#/definitions/ProductSpecification"
      relatedParty:
        description: ""
        type: "array"
        items:
          $ref: "#/definitions/RelatedParty"
      geographicAddress:
        description: ""
        type: "array"
        items:
          $ref: "#/definitions/GeographicAddress"
      geographicLocation:
        description: ""
        type: "array"
        items:
          $ref: "#/definitions/GeographicLocation"
      site:
        description: ""
        type: "array"
        items:
          $ref: "#/definitions/Site"
      productRelationship:
        description: ""
        type: "array"
        items:
          $ref: "#/definitions/ProductRelationship"
  OrderItemRelationShip:
    description: ""
    required:
    - "type"
    - "id"
    type: "object"
    properties:
      type:
        description: ""
        $ref: "#/definitions/RelationShipType"
      id:
        description: ""
        type: "string"
...
  RelatedQualification:
    description: ""
    type: "object"
    properties:
      id:
        description: ""
        type: "string"
      qualificationItem:
        description: ""
        type: "string"
  RelatedParty:
    description: ""
    required:
    - "name"
    - "role"
    - "number"
    type: "object"
    properties:
      id:
        description: "Unique identifier of a related party."
        type: "string"
      name:
        description: "Name of the related party."
        type: "string"
      role:
        description: "MEF: could be \"Buyer\", \"Seller\", \"Billing Contact\", \"\
          Order Contact\", \"Implementation Contact\", \"Technical Contact\", \"UNI\
          \ Site Contact\", \"UNI Alt Site Contact\", \"ENNI Site Contact\", \"ENNI\
          \ Alt Site Contact\""
        type: "string"
      numberExtension:
        description: "MEF Extension"
        type: "string"
      number:
        description: "Phone number  of the related party"
        type: "string"
      emailAddress:
        description: "Email address of the related party"
        type: "string"
  ProductOrderCreateIN:
    description: ""
    required:
    - "orderItem"
    type: "object"
    properties:
      externalId:
        description: ""
        type: "string"
      orderDate:
        description: ""
        type: "string"
        format: "date-time"
      requestedCompletionDate:
        description: ""
        type: "string"
        format: "date-time"
      requestedStartDate:
        description: ""
        type: "string"
        format: "date-time"
      expeditePriority:
        description: ""
        type: "integer"
        format: "int32"
      '@type':
        description: ""
        type: "string"
      '@schemaLocation':
        description: ""
        type: "string"
      orderActivity:
        description: ""
        $ref: "#/definitions/OrderActivity"
      desiredResponse:
        description: ""
        $ref: "#/definitions/desiredResponse"
      version:
        description: ""
        type: "string"
      projectId:
        description: ""
        type: "string"
      tspRestorationPriority:
        description: ""
        type: "string"
      relatedParty:
        description: ""
        type: "array"
        items:
          $ref: "#/definitions/RelatedParty"
      note:
        description: ""
        $ref: "#/definitions/Note"
      orderItem:
        description: ""
        type: "array"
        items:
          $ref: "#/definitions/OrderItemCreateIn"
...
  ProductSpecification:
    description: ""
    type: "object"
    properties:
      id:
        description: ""
        type: "string"
      describing:
        description: ""
        $ref: "#/definitions/Describing"
...
  GeographicLocation:
    description: "A GeographicLocation allows describing through coordinate(s) a point,\
      \ a line or a space"
    required:
    - "role"
    - "type"
    type: "object"
    properties:
      role:
        description: ""
        type: "string"
      type:
        description: ""
        type: "string"
      geographicPoint:
        description: "A GeoPoint defines a geographic point through coordinates."
        $ref: "#/definitions/GeographicPoint"
...
  OrderItemCreateIn:
    description: ""
    required:
    - "id"
    - "action"
    - "productOffering"
    type: "object"
    properties:
      id:
        description: "Identifier of the line item (generally it is a sequence number\
          \ 01, 02, 03, ...)"
        type: "string"
      action:
        description: "Can be \"add\" / \"modify\" / \"no_change\"/ \"delete\""
        $ref: "#/definitions/ActionType"
      billingAccount:
        description: "Identifier of the billing account where to bill the charges"
        $ref: "#/definitions/BillingAccountRef"
      productOffering:
        description: "A product offering reference (ProductOfferingRef). A product\
          \ offering represents entities that are orderable from the provider of the\
          \ catalog, this resource includes pricing information"
        $ref: "#/definitions/ProductOffering"
      product:
        description: ""
        $ref: "#/definitions/Product"
      orderItemRelationship:
        description: ""
        type: "array"
        items:
          $ref: "#/definitions/OrderItemRelationShip"
      qualificationRef:
        description: "Used to refer serviceability request done previously."
        type: "array"
        items:
          $ref: "#/definitions/RelatedQualification"
      '@type':
        description: ""
        type: "string"
      '@schemaLocation':
        description: ""
        type: "string"
      pricingMethod:
        description: "The pricing authority to be used to price the products on this\
          \ order. Choice of Tariff, Contract, Individual Case Basis (ICB), Other"
        $ref: "#/definitions/pricingMethod"
      pricingReference:
        description: "The value associated with the Pricing Method selected (e.g.\
          \ a contract or tariff id)"
        type: "string"
      pricingTerm:
        description: "The length of the contract in months"
        type: "integer"
        format: "int32"
  ProductOrderFindResp:
    description: ""
    required:
    - "externalId"
    - "orderDate"
    - "requestedCompletionDate"
    type: "object"
    properties:
      id:
        description: "ID created on repository side (OM system)\nMEF: Unique identifier\
          \ for the order that is generated by the Seller when the order is initially\
          \ accepted via an API."
        type: "string"
      externalId:
        description: "ID given by the consumer and only understandable by him (to\
          \ facilitate his searches afterwards).\nMEF: Buyer Purchase Order Number"
        type: "string"
      orderDate:
        description: "Date when the order was created"
        type: "string"
        format: "date-time"
      state:
        description: "Tracks the lifecycle status of the product order, such as Acknowledged,\
          \ Rejected, InProgress, Pending and so on\nMEF: to be seen if Desired Response\
          \ could be managed through state (9)"
        $ref: "#/definitions/StateType"
      requestedCompletionDate:
        description: "Requested delivery date from the requestor perspective"
        type: "string"
        format: "date-time"
      projectId:
        description: "An identifier that is used to group Orders that represent a\
          \ unit of functionality that is important to a Buyer. A Project can be used\
          \ to relate multiple Orders together."
        type: "string"